+++ src/test/java/com/rox/emu/rom/InesRomTest.java
@@ -11,7 +11,7 @@ import static org.junit.Assert.assertTrue;
 public class InesRomTest {
     @Test
     public void testFromWithBytes(){
-        final InesRom rom = InesRom.from(asPaddedHeader(new int[] {0x4E, 0x45, 0x53, 0x1A, 0x0, 0x0}));
+        final InesRom rom = InesRom.from(asPaddedHeader(new byte[] {0x4E, 0x45, 0x53, 0x1A, 0x0, 0x0}));
 
         assertNotNull(rom);
         assertFalse(rom.getDescription().isEmpty());
@@ -21,9 +21,9 @@ public class InesRomTest {
     public void testGetHeader(){
         final int prgRomBlocks = 0x9;
         final int chrRomBlocks = 0x4; // ChrRom/VRom
-        final int controlOptionsByte = 0b10101010;
+        final byte controlOptionsByte = (byte)0b10101010;
 
-        final InesRom rom = InesRom.from(asPaddedHeader(new int[] {0x4E, 0x45, 0x53, 0x1A, prgRomBlocks, chrRomBlocks, controlOptionsByte}));
+        final InesRom rom = InesRom.from(asPaddedHeader(new byte[] {0x4E, 0x45, 0x53, 0x1A, prgRomBlocks, chrRomBlocks, controlOptionsByte}));
         final InesRomHeader header = rom.getHeader();
 
         assertNotNull(header);
@@ -43,7 +43,7 @@ public class InesRomTest {
     @Test
     public void testInvalidFromWithBytes(){
         try {
-            InesRom.from(asPaddedHeader(new int[]{0x0, 0x0, 0x0, 0x1A, 0x0, 0x0}));
+            InesRom.from(asPaddedHeader(new byte[]{0x0, 0x0, 0x0, 0x1A, 0x0, 0x0}));
             fail("ROM contains an invalid NES header, should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -54,7 +54,7 @@ public class InesRomTest {
     public void testHeaderTooShort(){
         try {
 
-            InesRom.from(new int[]{'N', 'E', 'S'});
+            InesRom.from(new byte[]{'N', 'E', 'S'});
             fail("ROM contains a NES header that is too short to be valid, should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -64,7 +64,7 @@ public class InesRomTest {
     @Test
     public void testEmptyFrom(){
         try {
-            InesRom.from(new int[]{});
+            InesRom.from(new byte[]{});
             fail("ROM contains a missing NES header, should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -74,7 +74,7 @@ public class InesRomTest {
     @Test
     public void testInvalidFirstHeaderPrefixValue(){
         try {
-            InesRom.from(asPaddedHeader(new int[]{'Z', 'E', 'S', 0x1A, 0x0, 0x0}));
+            InesRom.from(asPaddedHeader(new byte[]{'Z', 'E', 'S', 0x1A, 0x0, 0x0}));
             fail("ROM contains an invalid NES header (first byte should be 'N'), should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -84,7 +84,7 @@ public class InesRomTest {
     @Test
     public void testInvalidSecondHeaderPrefixValue(){
         try {
-            InesRom.from(asPaddedHeader(new int[]{'N', 'Z', 'S', 0x1A, 0x0, 0x0}));
+            InesRom.from(asPaddedHeader(new byte[]{'N', 'Z', 'S', 0x1A, 0x0, 0x0}));
             fail("ROM contains an invalid NES header (second byte should be 0x45), should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -94,7 +94,7 @@ public class InesRomTest {
     @Test
     public void testInvalidThirdHeaderPrefixValue(){
         try {
-            InesRom.from(asPaddedHeader(new int[]{'N', 'E', 'Z', 0x1A, 0x0, 0x0}));
+            InesRom.from(asPaddedHeader(new byte[]{'N', 'E', 'Z', 0x1A, 0x0, 0x0}));
             fail("ROM contains an invalid NES header (third byte should be 0x53), should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
@@ -104,15 +104,15 @@ public class InesRomTest {
     @Test
     public void testInvalidFourthHeaderPrefixValue(){
         try {
-            InesRom.from(asPaddedHeader(new int[]{'N', 'E', 'S', 0x1B, 0x0, 0x0}));
+            InesRom.from(asPaddedHeader(new byte[]{'N', 'E', 'S', 0x1B, 0x0, 0x0}));
             fail("ROM contains an invalid NES header (fourth byte should be 0x1A), should throw an exception");
         }catch(UnknownRomException e){
             assertNotNull(e.getMessage());
         }
     }
 
-    private int[] asPaddedHeader(int[] values){
-        int[] header = new int[InesRomHeader.HEADER_SIZE];
+    private byte[] asPaddedHeader(byte[] values){
+        byte[] header = new byte[InesRomHeader.HEADER_SIZE];
         System.arraycopy(values, 0, header, 0, values.length);
         return header;
     }
